Added new timer, new flag, and encode edge as well as poly index in neighbour data

+++ DebugUtils/Include/DebugDrawLargeWorldCoordinates.h
@@ -6,10 +6,10 @@
 
 #if DU_LARGE_WORLD_COORDINATES_ENABLED
 
-typedef float duReal;
+typedef double duReal;
 
 #else // DU_LARGE_WORLD_COORDINATES_ENABLED
 
-typedef double duReal;
+typedef float duReal;
 
 #endif // DU_LARGE_WORLD_COORDINATES_ENABLED
